reincarnation commit
`magit-stgit.el' is no longer part of the Magit repository but
its karma is good and so it is reborn as a repository of its own.

To help it remember its past life do this:

  git remote add magit git://github.com/magit/magit.git
  git config remote.magit.tagopt --no-tags
  git fetch magit
  git replace --graft <this> 0a951ac7c7f152387002f75ebde1516cc7a506cb

See git-replace(1) for more information.

While it was still part of the Magit repository `magit-stgit.el'
was edited by the following people:

    99  Jonas Bernoulli
    16  Lluís Vilanova
     1  Peter J. Weisberg
     1  Pieter Praet
     2  Rüdiger Sonderfeld
     5  Yann Hodique