Revert "Add ignore file"
This reverts commit ed6f445ad883b5eb7cc75137e563cb856dc23f73.

Please don't do that - it is unnecessary and adds noise in the form of
an avoidable file.  This ignore rule belongs in a global ignore file.

There are two options (a) doing it in each and every repository
containing Emacs lisp code or (b) doing it once globally.  I very
strongly favor the second approach.