reincarnation commit
`magit-stgit.el' is no longer part of the Magit repository but
its karma is good and so it is reborn as a repository of its own.

To help it remember its past life do this:

  git remote add magit git://github.com/magit/magit.git
  git config remote.magit.tagopt --no-tags
  git fetch magit
  git replace --graft <this> 1176825ca145e35823000564d3b0bedccd334d05

See git-replace(1) for more information.

While it was still part of the Magit repository `magit-stgit.el'
was edited by the following people:

    68  Jonas Bernoulli
    16  Lluís Vilanova
     5  Yann Hodique
     2  Rüdiger Sonderfeld
     1  Peter J. Weisberg
     1  Pieter Praet