Directly depend on magit-popup
`magit' uses `transient' instead of `magit-popup' now.  Because
`magit-stgit' still uses `magit-popup' it has to depend on that
itself.

This makes it necessary to bump the `magit' dependency too.
`magit-popup' was split from `magit' in 2.12.0, so we have depend
on that version of both packages.

The dependency specification should be:

    ((emacs "24.4")
     (or ((magit "2.1.0 <= v < " 2.12.0"))
         ((magit "2.12.0")
          (magit-popup "2.12.0"))))

But `package.el' does not support that.